% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ showimage }
\alias{ showimage }
\alias{ showimage }
\title{ Show an Image on an R Graphics Device }
\description{
sometimes it is handy to be able to view an image file on an R graphics device.
}
