% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{show_image}
\alias{show_image}
\title{Show an Image on an R Graphics Device}
\usage{
show_image(file)
}
\arguments{
\item{file}{Name of the image file to show.}
}
\description{
Show an Image on an R Graphics Device
}

