% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showimage-package.R
\docType{package}
\name{showimage-package}
\alias{showimage}
\alias{showimage-package}
\title{showimage: Show an Image on an 'R' Graphics Device}
\description{
Sometimes it is handy to be able to view an image file on an 'R' graphics device. This package just does that. Currently it supports 'PNG' files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/showimage#readme}
  \item \url{https://r-lib.github.io/showimage/}
  \item Report bugs at \url{https://github.com/r-lib/showimage/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
